리팩토링: AppsInTossMenu 배포/브라우저 실행 로직 분리
- Editor/Menu/AITBrowserLauncher.cs: GetBrowserPath(ServerType) 및
  OpenBrowser(int, ServerType) 추출. LaunchServerProcess의 인라인
  WaitForPortAndOpenBrowser 호출을 AITBrowserLauncher.OpenBrowser로 위임.

- Editor/Menu/AITDeployManager.cs: 배포 클러스터 전체 이전.
  _buildEntryInProgress 재진입 가드, RunPublish, RunBuildAndPackage,
  ExecuteDeploy, GetPhaseText, ExtractDeployUrl, ExtractDeployErrorMessage,
  ShowBuildFailedDialog 포함. 독립 _buildStopwatch 사용 (StartServer의
  buildStopwatch와 분리).

- Editor/Menu/DeploySuccessWindow.cs: 중첩 EditorWindow를 독립 파일로
  분리. namespace AppsInToss.Editor.Menu, internal class로 변경.

- Editor/AppsInTossMenu.cs: BuildAndPackage/Publish를 얇은 위임자로
  축소. StartServer의 ShowBuildFailedDialog 호출을
  AITDeployManager.ShowBuildFailedDialog로 전환. 배포/브라우저/DeploySuccessWindow
  코드 제거 (~640줄 감소).

- TODO.md: P1 AppsInTossMenu 분리 항목 제거.
