+++ Editor/Menu/AITBrowserLauncher.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d006a3e1859b450281c5e9f65d1d4fa3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
