+++ Editor/Menu/AITDeployManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa214fb8f4004d5ba76c54bdd3b0d925
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
