+++ Editor/Menu/DeploySuccessWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d357039c00944d55816bec1ebbe1e94b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
